requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf2
python-docx
pandas
pyyaml
groq
